+++ Moco/Moco/MocoUtilities.h
@@ -87,6 +87,15 @@ inline bool endsWith(const std::string& string, const std::string& ending) {
     return false;
 }
 
+/// An OpenSim XML file may contain file paths that are relative to the
+/// directory containing the XML file; use this function to convert that
+/// relative path into an absolute path.
+/// @ingroup mocogenutil
+OSIMMOCO_API
+std::string getAbsolutePathnameFromXMLDocument(
+        const std::string& documentFileName,
+        const std::string& pathnameRelativeToDocument);
+
 /// @name Filling in a string with variables.
 /// @{
 
@@ -813,11 +822,11 @@ private:
 };
 
 /// Obtain the ground reaction forces, centers of pressure, and torques
-/// resulting from Force elements (e.g., SmoothSphereHalfSpaceForce), using the
-/// model and the trajectory. Forces and torques are expressed in the ground
+/// resulting from Force elements (e.g., SmoothSphereHalfSpaceForce), using a
+/// model and states trajectory. Forces and torques are expressed in the ground
 /// frame with respect to the ground origin. Hence, the centers of pressure are
-/// at the origin. Names of Force elements should be provided separately for
-/// elements of the right and left feet. The output is a table formated for use
+/// at the origin. Paths to Force elements should be provided separately for
+/// elements of the right and left feet. The output is a table formatted for use
 /// with OpenSim tools; the labels of the columns distinguish between right
 /// ("<>_r") and left ("<>_l") forces, centers of pressure, and torques. The
 /// forces and torques used are taken from the first six outputs of
@@ -827,9 +836,17 @@ private:
 /// @ingroup mocomodelutil
 OSIMMOCO_API
 TimeSeriesTable createExternalLoadsTableForGait(Model model,
+        const StatesTrajectory& trajectory,
+        const std::vector<std::string>& forcePathsRightFoot,
+        const std::vector<std::string>& forcePathsLeftFoot);
+
+/// Same as above, but with a MocoTrajectory instead of a StatesTrajectory.
+/// @ingroup mocomodelutil
+OSIMMOCO_API
+TimeSeriesTable createExternalLoadsTableForGait(Model model,
         const MocoTrajectory& trajectory,
-        const std::vector<std::string>& forceNamesRightFoot,
-        const std::vector<std::string>& forceNamesLeftFoot);
+        const std::vector<std::string>& forcePathsRightFoot,
+        const std::vector<std::string>& forcePathsLeftFoot);
 
 /// Solve for the root of a scalar function using the bisection method.
 /// @param calcResidual a function that computes the error
@@ -842,7 +859,7 @@ TimeSeriesTable createExternalLoadsTableForGait(Model model,
 OSIMMOCO_API
 SimTK::Real solveBisection(
         std::function<double(const double&)> calcResidual,
-        double left, double right, const double& xTolerance = 1e-6,
+        double left, double right, const double& tolerance = 1e-6,
         int maxIterations = 1000);
 
 } // namespace OpenSim