+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 Change Log
 ==========
-
-0.3.0 (in development)
+0.4.0 (in development) 
 ----------------------
+- 2020-01-28: Added a projection setting to MocoFrameDistanceConstraint.
+
+0.3.0 
+-----
+- 2020-01-15: MocoContactTrackingGoal allows minimizing the error between
+              compliant contact forces and experimentally measured contact 
+              forces.
+
 - 2020-01-10: MocoSolution now provides a breakdown of terms in the cost 
               (only when using MocoCasADiSolver).